Move filtering code from PlateLoader to new FilterHub
